fix/Types refactoring pages
